Create py.listcomprehension
This is list comprehension is use for list variables define and this code is also use for list of any thing to changes variables there in list and both the code uses in this list variables changes. I given both the code to help for in list comprehension and this your choice what to do use.