Create py.read excel data in python
Using this to find easily your excel files data in python and check the rows and columns value in sheets. This code in location of file is blank because in this blank area to you added your file location and type all code then you find your excel file data in python.